requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai==1.72.0
python-dotenv==0.19.0
pydantic==2.6.1
jinja2==3.1.6
pytest==8.0.0
httpx==0.26.0